@@ -54,22 +51,30 @@
     "three.meshline": "^1.2.0",
     "tinyqueue": "^2.0.3",
     "umap-js": "^1.3.2",
-    "vega": "^5.17.0",
-    "vega-lite": "^4.16.8",
-    "vega-tooltip": "^0.24.2"
+    "url-loader": "^4.1.1",
+    "vega": "5.17.0",
+    "vega-lite": "4.16.8",
+    "vega-typings": "0.19.1",
+    "worker-loader": "^3.0.8"
   },
   "devDependencies": {
     "@babel/core": "^7.10.2",
     "@babel/preset-env": "^7.10.2",
     "@babel/preset-react": "^7.8.3",
     "@babel/preset-typescript": "^7.8.3",
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
+    "@mui/material": "^5.0.6",
+    "@mui/styles": "^5.0.2",
     "@svgr/webpack": "^5.5.0",
-    "@types/react": "^16.9.19",
-    "@types/react-dom": "^16.9.5",
-    "babel-loader": "^8.0.6",
+    "babel-loader": "^8.2.3",
     "css-loader": "^3.4.2",
     "jest": "^26.6.3",
     "node-sass": "^4.13.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
+    "redux": "^4.1.2",
     "sass-loader": "^8.0.2",
     "shader-loader": "^1.3.1",
     "source-map-loader": "^0.2.4",
@@ -77,9 +82,19 @@
     "svg-url-loader": "^7.1.1",
     "ts-loader": "^6.2.1",
     "typescript": "^3.7.5",
-    "webpack": "^4.41.5",
+    "webpack": "^4.46.0",
     "webpack-cli": "^3.3.10",
     "webpack-glsl-loader": "^1.0.1"
+  },
+  "peerDependencies": {
+    "@mui/styles": "^5.0.1",
+    "@emotion/styled": "^11.3.0",
+    "@emotion/react": "^11.5.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.0",
+    "redux": "^4.0.5",
+    "@mui/material": "^5.0.3"
   },
   "scripts": {
     "webpack:dev": "npx webpack --config webpack.config.js",
