@@ -1,55 +1,53 @@
-
 import { getDefaultZoom, arraysEqual, normalizeWheel, centerOfMass, interpolateLinear, generateZoomForSet } from './UtilityFunctions';
 import { LassoSelection } from './tools'
 import * as React from "react";
 import * as THREE from 'three'
-import { LassoLayer } from './LassoLayer/LassoLayer';
-import Cluster from '../Utility/Data/Cluster';
+import { LassoLayer } from './LassoLayer';
+import { ACluster, ICluster, isCluster } from '../../model/Cluster';
+import { TypedObject } from "../../model/TypedObject";
 import { connect, ConnectedProps } from 'react-redux'
-import { Tool, getToolCursor } from '../Overlays/ToolSelection/ToolSelection';
-import { Vect } from "../Utility/Data/Vect";
-import { setClusterEdgesAction } from "../Ducks/ClusterEdgesDuck";
-import { setViewTransform } from "../Ducks/ViewTransformDuck";
-import { aggSelectCluster, setAggregationAction, toggleAggregationAction } from "../Ducks/AggregationDuck";
+import { isVector, IVector } from "../../model/Vector";
+import { setViewTransform, ViewTransformType } from "../Ducks/ViewTransformDuck";
+import { selectClusters, selectVectors } from "../Ducks/AggregationDuck";
 import { CameraTransformations } from './CameraTransformations'
 import { Camera } from 'three';
 import { LineVisualization, PointVisualization } from './meshes';
-import { MultivariateClustering } from './Visualizations/MultivariateClustering';
+import { MultivariateClustering } from './MultivariateClustering';
 import { DisplayMode, displayModeSupportsStates } from '../Ducks/DisplayModeDuck';
 import { setActiveLine } from '../Ducks/ActiveLineDuck';
 import { mappingFromScale } from '../Utility/Colors/colors';
 import { setPointColorMapping } from '../Ducks/PointColorMappingDuck';
 import { RootState } from '../Store/Store';
-import { Menu, MenuItem } from '@material-ui/core';
+import { Menu, MenuItem } from '@mui/material';
 import * as nt from '../NumTs/NumTs'
 import { MouseController } from './MouseController';
-import { addClusterToStory, addEdgeToActive, addStory, removeClusterFromStories, removeEdgeFromActive, setActiveStory, setActiveTrace } from '../Ducks/StoriesDuck';
-import { setLineUpInput_visibility, setLineUpInput_dump, setLineUpInput_filter } from '../Ducks/LineUpInputDuck';
-import { Storybook } from '../Utility/Data/Storybook';
+import { addCluster, addEdgeToActive, addBook, deleteCluster, removeEdgeFromActive, setActiveTrace, AStorytelling } from '../Ducks/StoriesDuck';
+import { IBook, ABook } from '../../model/Book';
 import { RenderingContextEx } from '../Utility/RenderingContextEx';
-import { Edge } from '../Utility/graphs';
+import { IEdge, isEdge } from "../../model/Edge";
 import { getSyncNodesAlt } from '../NumTs/NumTs';
-import { ClusterDragTool } from './Tools/ClusterDragTool';
-import { TraceSelectTool } from './Tools/TraceSelectTool';
-import { Embedding } from '../Utility/Data/Embedding';
+import { ClusterDragTool } from './ClusterDragTool';
+import { TraceSelectTool } from './TraceSelectTool';
 import { setOpenTabAction } from '../Ducks/OpenTabDuck';
 import { setHoverState } from '../Ducks/HoverStateDuck';
-import './WebGl.scss'
 import { pointInHull } from '../Utility/Geometry/Intersection';
+import { ADataset } from '../../model/Dataset';
+import { DataLine } from '../../model/DataLine';
+import { ObjectTypes } from '../../model/ObjectType';
+import { v4 as uuidv4 } from 'uuid';
+import { ComponentConfig } from '../../Application';
+import { ANormalized } from '../Utility/NormalizedState';
 
 type ViewState = {
-    displayClusters: any
     camera: Camera
     menuX: number
     menuY: number
-    menuTarget: any
+    menuTarget: TypedObject
 }
 
 const mapStateToProps = (state: RootState) => ({
-    currentTool: state.currentTool,
     currentAggregation: state.currentAggregation,
     vectorByShape: state.vectorByShape,
-    checkedShapes: state.checkedShapes,
     dataset: state.dataset,
     highlightedSequence: state.highlightedSequence,
     activeLine: state.activeLine,
@@ -63,48 +61,39 @@ const mapStateToProps = (state: RootState) => ({
     channelSize: state.channelSize,
     channelColor: state.channelColor,
     channelBrightness: state.channelBrightness,
-    pointColorScale: state.pointColorScale,
     stories: state.stories,
     trailSettings: state.trailSettings,
+    hoverState: state.hoverState,
+    workspace: state.projections.workspace,
+    colorScales: state.colorScales,
+    pointDisplay: state.pointDisplay
 })
-
 
 const mapDispatchToProps = dispatch => ({
-    setCurrentAggregation: id => dispatch(setAggregationAction(id)),
-    setClusterEdges: clusterEdges => dispatch(setClusterEdgesAction(clusterEdges)),
+    selectVectors: (vectors: number[], shiftKey: boolean) => dispatch(selectVectors(vectors, shiftKey)),
     setActiveLine: activeLine => dispatch(setActiveLine(activeLine)),
     setViewTransform: (camera, width, height) => dispatch(setViewTransform(camera, width, height)),
-    toggleAggregation: aggregation => dispatch(toggleAggregationAction(aggregation)),
     setHoverState: (hoverState, updater) => dispatch(setHoverState(hoverState, updater)),
     setPointColorMapping: mapping => dispatch(setPointColorMapping(mapping)),
-    removeClusterFromStories: cluster => dispatch(removeClusterFromStories(cluster)),
-    // setSelectedClusters: clusters => dispatch(setSelectedClusters(clusters)),
-    // setLineUpInput_data: input => dispatch(setLineUpInput_data(input)),
-    // setLineUpInput_columns: input => dispatch(setLineUpInput_columns(input)),
-    setLineUpInput_visibility: input => dispatch(setLineUpInput_visibility(input)),
-    setLineUpInput_dump: input => dispatch(setLineUpInput_dump(input)),
-    setLineUpInput_filter: input => dispatch(setLineUpInput_filter(input)),
-    addStory: story => dispatch(addStory(story)),
-    addClusterToStory: cluster => dispatch(addClusterToStory(cluster)),
-    setActiveStory: story => dispatch(setActiveStory(story)),
-    addEdgeToActive: edge => dispatch(addEdgeToActive(edge)),
+    removeClusterFromStories: (cluster: ICluster) => dispatch(deleteCluster(cluster)),
+    addStory: (story: IBook, activate: boolean) => dispatch(addBook(story, activate)),
+    addClusterToStory: cluster => dispatch(addCluster(cluster)),
+    addEdgeToActive: (edge: IEdge) => dispatch(addEdgeToActive(edge)),
     setActiveTrace: trace => dispatch(setActiveTrace(trace)),
     setOpenTab: tab => dispatch(setOpenTabAction(tab)),
-    setSelectedCluster: (cluster, shiftKey) => dispatch(aggSelectCluster(cluster, shiftKey)),
+    setSelectedCluster: (clusters: string[], shiftKey: boolean) => dispatch(selectClusters(clusters, shiftKey)),
     removeEdgeFromActive: (edge) => dispatch(removeEdgeFromActive(edge))
 })
-
 
 const connector = connect(mapStateToProps, mapDispatchToProps, null, { forwardRef: true });
 
 type PropsFromRedux = ConnectedProps<typeof connector>
 
-type Props = PropsFromRedux
-
-
+type Props = PropsFromRedux & {
+    overrideComponents: ComponentConfig
+}
 
 const UPDATER = "scatter";
-
 
 export const WebGLView = connector(class extends React.Component<Props, ViewState> {
     lasso: LassoSelection
@@ -118,16 +107,15 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
     mouse: any
     mouseDownPosition: any
     initialMousePosition: any;
-    currentHover: any;
-    currentAggregation: Vect[];
+    currentHover: TypedObject;
     camera: any;
-    vectors: Vect[];
+    vectors: IVector[];
     renderer: any;
     lines: LineVisualization;
     scene: THREE.Scene;
     dataset: any;
     lineColorScheme: any;
-    segments: any;
+    segments: DataLine[];
     pointScene: THREE.Scene;
     vectorColorScheme: any;
     prevTime: number;
@@ -166,7 +154,6 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
         this.state = {
             // clusters to display using force-directed layout
-            displayClusters: [],
             camera: null,
             menuX: null,
             menuY: null,
@@ -174,41 +161,6 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         }
 
     }
-
-
-    lineupFilterUpdate() {
-        this.particles.update()
-        this.requestRender()
-    }
-
-    async hoverUpdate(hover_item, updater) {
-        let idx = -1;
-        if (hover_item && hover_item["__meta__"]) {
-            idx = hover_item["__meta__"]["view"]["meshIndex"];
-        }
-        this.particles.highlight(idx);
-        if (this.props.dataset.isSequential) {
-            if (idx >= 0) {
-                this.lines.highlight([this.props.dataset.vectors[idx].view.segment.lineKey], this.getWidth(), this.getHeight(), this.scene)
-            } else {
-                this.lines.highlight([], this.getWidth(), this.getHeight(), this.scene)
-            }
-        }
-        if (idx >= 0) {
-            if (this.currentHover != this.props.dataset.vectors[idx]) {
-                this.currentHover = this.props.dataset.vectors[idx]
-                this.props.setHoverState(this.props.dataset.vectors[idx], updater)
-                this.requestRender()
-            }
-        } else {
-            if (this.currentHover) {
-                this.currentHover = null
-                this.props.setHoverState(null, updater)
-                this.requestRender()
-            }
-        }
-    }
-
 
     /**
      * Initializes the callbacks for the MouseController.
@@ -236,21 +188,17 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         this.mouseController.onDragStart = (event: MouseEvent, button: number, initial: nt.VectorType) => {
             switch (button) {
                 case 0:
-                    switch (this.props.currentTool) {
-                        case Tool.Default:
-                            // Check if we have dragged from a cluster
-                            let cluster = this.chooseCluster(initial)
-                            if (cluster) {
-                                // Initiate action to let user drag an arrow to other cluster
-                                this.clusterDrag = new ClusterDragTool(cluster)
-                            } else {
-                                // Initiate lasso selection
-                                let initialWorld = CameraTransformations.screenToWorld(initial, this.createTransform())
+                    // Check if we have dragged from a cluster
+                    let cluster = this.chooseCluster(initial)
+                    if (cluster) {
+                        // Initiate action to let user drag an arrow to other cluster
+                        this.clusterDrag = new ClusterDragTool(cluster)
+                    } else {
+                        // Initiate lasso selection
+                        let initialWorld = CameraTransformations.screenToWorld(initial, this.createTransform())
 
-                                this.lasso = new LassoSelection()
-                                this.lasso.mouseDown(true, initialWorld.x, initialWorld.y)
-                            }
-                            break;
+                        this.lasso = new LassoSelection()
+                        this.lasso.mouseDown(true, initialWorld.x, initialWorld.y)
                     }
                     break;
             }
@@ -260,46 +208,48 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         this.mouseController.onDragEnd = (event: MouseEvent, button: number) => {
             switch (button) {
                 case 0:
-                    switch (this.props.currentTool) {
-                        case Tool.Default:
-                            let coords = CameraTransformations.screenToWorld({ x: event.offsetX, y: event.offsetY }, this.createTransform())
+                    let coords = CameraTransformations.screenToWorld({ x: event.offsetX, y: event.offsetY }, this.createTransform())
 
-                            // In case we dragged a cluster... connect them
-                            if (this.clusterDrag) {
-                                let cluster = this.chooseCluster({ x: event.offsetX, y: event.offsetY })
-                                if (cluster) {
-                                    this.props.addEdgeToActive(new Edge(this.clusterDrag.cluster, cluster, null))
-                                }
+                    // In case we dragged a cluster... connect them
+                    if (this.clusterDrag) {
+                        let cluster = this.chooseCluster({ x: event.offsetX, y: event.offsetY })
 
-                                this.clusterDrag = null
+                        if (cluster) {
+                            const activeStory = AStorytelling.getActive(this.props.stories)
+
+                            const source = Object.keys(activeStory.clusters.byId).find(key => activeStory.clusters.byId[key] === this.clusterDrag.cluster)
+                            const destination = Object.keys(activeStory.clusters.byId).find(key => activeStory.clusters.byId[key] === cluster)
+
+                            if (ANormalized.entries(activeStory.edges).find(e => e[1].source === source && e[1].destination === destination) !== undefined) {
+                                this.props.addEdgeToActive({
+                                    source,
+                                    destination,
+                                    objectType: ObjectTypes.Edge
+                                })
                             }
+                        }
 
-                            // In case we have a lasso, select states
-                            if (this.lasso) {
-                                // If there is an active lasso, process it
-                                var wasDrawing = this.lasso.drawing
+                        this.clusterDrag = null
+                    }
 
-                                this.lasso.mouseUp(coords.x, coords.y)
+                    // In case we have a lasso, select states
+                    if (this.lasso) {
+                        // If there is an active lasso, process it
+                        var wasDrawing = this.lasso.drawing
 
-                                var indices = this.lasso.selection(this.props.dataset.vectors, (vector) => this.particles.isPointVisible(vector))
-                                if (indices.length > 0 && wasDrawing && displayModeSupportsStates(this.props.displayMode)) {
-                                    var selected = indices.map(index => this.props.dataset.vectors[index])
+                        this.lasso.mouseUp(coords.x, coords.y)
 
-                                    if (event.ctrlKey) {
-                                        this.props.toggleAggregation(selected)
-                                    } else {
-                                        this.props.setCurrentAggregation(selected)
-                                    }
+                        var indices = this.lasso.selection(this.props.dataset, this.props.workspace, (vector) => this.particles.isPointVisible(vector))
+                        if (indices.length > 0 && wasDrawing && displayModeSupportsStates(this.props.displayMode)) {
+                            this.props.selectVectors(indices, event.ctrlKey)
 
-                                    this.props.setOpenTab(4)
+                            this.props.setOpenTab(4)
 
-                                } else if (wasDrawing) {
-                                    this.clearSelection()
-                                }
+                        } else if (wasDrawing) {
+                            this.clearSelection()
+                        }
 
-                                this.lasso = null
-                            }
-                            break;
+                        this.lasso = null
                     }
                     break;
             }
@@ -309,21 +259,8 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
             switch (button) {
                 case 0:
-                    switch (this.props.currentTool) {
-                        case Tool.Default:
-                            let coords = CameraTransformations.screenToWorld({ x: event.offsetX, y: event.offsetY }, this.createTransform())
-                            this.lasso?.mouseMove(coords.x, coords.y)
-                            break;
-                        case Tool.Move:
-                            this.camera.position.x = this.camera.position.x - CameraTransformations.pixelToWorldCoordinates(event.movementX, this.createTransform())
-                            this.camera.position.y = this.camera.position.y + CameraTransformations.pixelToWorldCoordinates(event.movementY, this.createTransform())
-
-                            this.camera.updateProjectionMatrix()
-                            this.props.setViewTransform(this.camera, this.getWidth(), this.getHeight())
-
-                            this.requestRender()
-                            break;
-                    }
+                    let coords = CameraTransformations.screenToWorld({ x: event.offsetX, y: event.offsetY }, this.createTransform())
+                    this.lasso?.mouseMove(coords.x, coords.y)
                     break;
                 case 2:
                     this.camera.position.x = this.camera.position.x - CameraTransformations.pixelToWorldCoordinates(event.movementX, this.createTransform())
@@ -338,90 +275,89 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         }
 
         this.mouseController.onContext = (event: MouseEvent, button: number) => {
-
-            if (this.props.currentTool == Tool.Default) {
-                switch (button) {
-                    case 0:
-                        if (this.props.activeLine) {
-                            break;
-                        }
-
-
-
-
-                        if (this.traceSelect) {
-                            let target = this.chooseCluster({ x: event.offsetX, y: event.offsetY })
-
-                            if (target) {
-                                // We want to select a trace between 2 clusters
-                                let paths = Storybook.depthFirstSearch(this.props.stories.active.toGraph(), this.traceSelect.cluster.label, target.label)
-
-                                if (paths.length > 0) {
-                                    let mainPath = paths[0].map(id => this.props.stories.active.clusters.find(e => e.label == id))
-                                    let mainEdges = mainPath.slice(1).map((item, index) => {
-                                        return this.props.stories.active.edges.find(edge => edge.source == mainPath[index] && edge.destination == item)
-                                    })
-                                    this.props.setActiveTrace({
-                                        mainPath: mainPath,
-                                        mainEdges: mainEdges,
-                                        sidePaths: paths.slice(1).map(ids => {
-                                            let path = ids.map(id => this.props.stories.active.clusters.find(e => e.label == id))
-                                            let edges = path.slice(1).map((item, index) => {
-                                                return this.props.stories.active.edges.find(edge => edge.source == path[index] && edge.destination == item)
-                                            })
-                                            return {
-                                                nodes: path,
-                                                edges: edges,
-                                                syncNodes: getSyncNodesAlt(mainPath, path)
-                                            }
-                                        })
-                                    })
-                                }
-                            }
-
-                            this.traceSelect = null
-                        } else if (this.currentHover && this.currentHover instanceof Vect) {
-                            // We click on a hover target
-                            if (event.ctrlKey) {
-                                // There is a hover target ... select it
-                                this.props.toggleAggregation([this.currentHover])
-                            } else {
-                                this.props.setCurrentAggregation([this.currentHover])
-                            }
-                        } else if (this.currentHover && this.currentHover instanceof Cluster) {
-                            this.props.setSelectedCluster(this.currentHover, event.ctrlKey)
-                        }
-
-
+            switch (button) {
+                case 0:
+                    if (this.props.activeLine) {
                         break;
-                    case 2:
-                        let cluster = this.chooseCluster({ x: event.offsetX, y: event.offsetY })
+                    }
 
-                        if (cluster) {
+
+
+
+                    if (this.traceSelect) {
+                        let target = this.chooseCluster({ x: event.offsetX, y: event.offsetY })
+
+                        if (target) {
+                            const activeStory = AStorytelling.getActive(this.props.stories)
+                            // We want to select a trace between 2 clusters
+                            let paths = ABook.depthFirstSearch(ABook.toGraph(activeStory),
+                                Object.entries(activeStory.clusters.byId).find(([key, value]) => value === this.traceSelect.cluster)[0],
+                                Object.entries(activeStory.clusters.byId).find(([key, value]) => value === target)[0])
+
+                            if (paths.length > 0) {
+                                let mainPath = paths[0]
+                                let mainEdges = mainPath.slice(1).map((item, index) => {
+                                    const [resultEdgeKey, _] = Object.entries(activeStory.edges.byId).find(([key, edge]) => edge.source === mainPath[index] && edge.destination === item)
+                                    return resultEdgeKey
+                                })
+                                this.props.setActiveTrace({
+                                    mainPath: mainPath,
+                                    mainEdges: mainEdges,
+                                    sidePaths: paths.slice(1).map(ids => {
+                                        let path = ids
+                                        let edges = path.slice(1).map((item, index) => {
+                                            const [resultEdgeKey, _] = Object.entries(activeStory.edges.byId).find(([key, edge]) => edge.source === path[index] && edge.destination === item)
+                                            return resultEdgeKey
+                                        })
+                                        return {
+                                            nodes: path,
+                                            edges: edges,
+                                            syncNodes: getSyncNodesAlt(mainPath, path)
+                                        }
+                                    })
+                                })
+                            }
+                        }
+
+                        this.traceSelect = null
+                    } else if (this.currentHover && isVector(this.currentHover)) {
+
+                        // We click on a hover target
+                        this.props.selectVectors([this.currentHover.__meta__.meshIndex], event.ctrlKey)
+                    } else if (this.currentHover && isCluster(this.currentHover)) {
+                        const activeStory = AStorytelling.getActive(this.props.stories)
+                        this.props.setSelectedCluster([Object.keys(activeStory.clusters.byId).find(key => activeStory.clusters.byId[key] === this.currentHover)], event.ctrlKey)
+                    }
+
+
+                    break;
+                case 2:
+                    let cluster = this.chooseCluster({ x: event.offsetX, y: event.offsetY })
+
+                    if (cluster) {
+                        this.setState({
+                            menuX: event.clientX,
+                            menuY: event.clientY,
+                            menuTarget: cluster
+                        })
+                    } else {
+                        let edge = this.chooseEdge(this.mouseController.currentMousePosition)
+                        if (edge) {
                             this.setState({
                                 menuX: event.clientX,
                                 menuY: event.clientY,
-                                menuTarget: cluster
+                                menuTarget: edge
                             })
                         } else {
-                            let edge = this.chooseEdge(this.mouseController.currentMousePosition)
-                            if (edge) {
-                                this.setState({
-                                    menuX: event.clientX,
-                                    menuY: event.clientY,
-                                    menuTarget: edge
-                                })
-                            } else {
-                                this.setState({
-                                    menuX: event.clientX,
-                                    menuY: event.clientY,
-                                    menuTarget: null
-                                })
-                            }
+                            this.setState({
+                                menuX: event.clientX,
+                                menuY: event.clientY,
+                                menuTarget: null
+                            })
                         }
+                    }
 
-                        break;
-                }
+                    break;
             }
         }
 
@@ -431,146 +367,104 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
 
         this.mouseController.onMouseMove = (event: MouseEvent) => {
-            switch (this.props.currentTool) {
-                case Tool.Default:
-                    if (this.props.displayMode == DisplayMode.OnlyClusters) {
-                        break;
+            if (this.props.displayMode == DisplayMode.OnlyClusters) {
+                return;
+            }
+
+            // In case we have a line in the sequence UI
+            if (this.props.activeLine) {
+                return;
+            }
+
+            if (this.infoTimeout != null) {
+                clearTimeout(this.infoTimeout)
+            }
+
+            this.infoTimeout = setTimeout(() => {
+                this.infoTimeout = null
+
+                let cluster = this.chooseCluster(this.mouseController.currentMousePosition)
+                if (cluster) {
+                    if (this.currentHover != cluster) {
+                        this.currentHover = cluster
+                        this.props.setHoverState(cluster, UPDATER)
+
+                        if (this.props.dataset.isSequential) {
+                            this.lines.highlight([], this.getWidth(), this.getHeight(), this.scene)
+                        }
+                        this.particles.highlight(-1)
+
+                        this.requestRender()
                     }
+                } else {
+                    let coords = CameraTransformations.screenToWorld(this.mouseController.currentMousePosition, this.createTransform())
 
-                    // In case we have a line in the sequence UI
-                    if (this.props.activeLine) {
-                        break;
-                    }
+                    let edge = this.chooseEdge(this.mouseController.currentMousePosition)
 
-                    if (this.infoTimeout != null) {
-                        clearTimeout(this.infoTimeout)
-                    }
 
-                    this.infoTimeout = setTimeout(() => {
-                        this.infoTimeout = null
 
-                        let cluster = this.chooseCluster(this.mouseController.currentMousePosition)
-                        if (cluster) {
-                            if (this.currentHover != cluster) {
-                                this.currentHover = cluster
-                                this.props.setHoverState(cluster, UPDATER)
+                    if (edge) {
+                        if (this.currentHover !== edge) {
+                            this.currentHover = edge
+                            this.props.setHoverState(edge, UPDATER)
+                        }
+                    } else {
+                        // Get index of selected node
+                        var idx = this.choose(coords)
+                        this.particles.highlight(idx)
 
-                                if (this.props.dataset.isSequential) {
-                                    this.lines.highlight([], this.getWidth(), this.getHeight(), this.scene)
-                                }
-                                this.particles.highlight(-1)
+                        if (this.props.dataset.isSequential) {
+                            if (idx >= 0) {
+                                this.lines.highlight([this.props.dataset.vectors[idx].line], this.getWidth(), this.getHeight(), this.scene)
+                            } else {
+                                this.lines.highlight([], this.getWidth(), this.getHeight(), this.scene)
+                            }
+                        }
 
+
+
+                        if (idx >= 0) {
+                            if (this.currentHover != this.props.dataset.vectors[idx]) {
+                                this.currentHover = this.props.dataset.vectors[idx]
+
+                                this.props.setHoverState(this.props.dataset.vectors[idx], UPDATER)
                                 this.requestRender()
                             }
                         } else {
-                            let coords = CameraTransformations.screenToWorld(this.mouseController.currentMousePosition, this.createTransform())
-
-                            let edge = this.chooseEdge(this.mouseController.currentMousePosition)
-                            if (edge) {
-                                console.log("edge")
-                                if (this.currentHover !== edge) {
-                                    this.currentHover = edge
-                                    this.props.setHoverState(edge, UPDATER)
-                                }
-                            } else {
-                                // Get index of selected node
-                                var idx = this.choose(coords)
-                                this.particles.highlight(idx)
-
-                                if (this.props.dataset.isSequential) {
-                                    if (idx >= 0) {
-                                        this.lines.highlight([this.props.dataset.vectors[idx].view.segment.lineKey], this.getWidth(), this.getHeight(), this.scene)
-                                    } else {
-                                        this.lines.highlight([], this.getWidth(), this.getHeight(), this.scene)
-                                    }
-                                }
-
-
-
-                                if (idx >= 0) {
-                                    if (this.currentHover != this.props.dataset.vectors[idx]) {
-                                        this.currentHover = this.props.dataset.vectors[idx]
-
-                                        this.props.setHoverState(this.props.dataset.vectors[idx], UPDATER)
-                                        this.requestRender()
-                                    }
-                                } else {
-                                    if (this.currentHover) {
-                                        this.currentHover = null
-                                        this.props.setHoverState(null, UPDATER)
-                                        this.requestRender()
-                                    }
-                                }
+                            if (this.currentHover) {
+                                this.currentHover = null
+                                this.props.setHoverState(null, UPDATER)
+                                this.requestRender()
                             }
                         }
-                    }, 10);
-
-                    break;
-            }
-
+                    }
+                }
+            }, 10);
         }
     }
 
 
 
 
-    /**
-     * Creates the line visualization for given attribute key, e. g. line or algo
-     * 
-     * @param attribute The attribute key
-     */
-    recreateLines(attribute: string) {
-        const dataset = this.props.dataset
-        const segments = dataset.getSegs(attribute)
-
-        dataset.segments = segments
-
-
-        this.lines?.dispose(this.scene)
-        this.lines = null
-
-        this.lines = new LineVisualization(segments, this.lineColorScheme)
-        this.lines.createMesh(this.props.lineBrightness)
-        this.lines.setZoom(this.camera.zoom)
-
-        // First add lines, then particles
-        this.lines.meshes.forEach(line => this.scene.add(line.line))
-
-        this.pointScene.remove(this.particles?.mesh)
-        this.particles = new PointVisualization(this.vectorColorScheme, this.props.dataset, window.devicePixelRatio * 8, this.lines?.grayedLayerSystem)
-        this.particles.createMesh(this.props.dataset.vectors, segments)
-        this.particles.zoom(this.camera.zoom)
-        this.particles.updateColor()
-        this.particles.update()
-
-
-        //this.scene.add(this.particles.mesh);
-        this.pointScene.add(this.particles.mesh)
-
-        this.requestRender()
-    }
-
-
-
-
-
-
-
-
-
-    chooseCluster(screenPosition: { x: number, y: number }) {
-        let nearest = null
+    chooseCluster(screenPosition: { x: number, y: number }): ICluster {
+        let nearest: ICluster = null
         let min = Number.MAX_SAFE_INTEGER
 
-        this.props.stories.active?.clusters.forEach(cluster => {
-            let clusterScreen = CameraTransformations.worldToScreen(new THREE.Vector2(cluster.getCenter().x, cluster.getCenter().y), this.createTransform())
+        const activeStory = AStorytelling.getActive(this.props.stories)
+
+        if (!activeStory) {
+            return null
+        }
+
+        for (const [key, cluster] of Object.entries(activeStory.clusters.byId)) {
+            let clusterScreen = CameraTransformations.worldToScreen(new THREE.Vector2(ACluster.getCenterFromWorkspace(this.props.workspace, cluster).x, ACluster.getCenterFromWorkspace(this.props.workspace, cluster).y), this.createTransform())
             let dist = nt.euclideanDistance(screenPosition.x, screenPosition.y, clusterScreen.x, clusterScreen.y)
 
             if (dist < min && dist < 16) {
                 nearest = cluster
                 min = dist
             }
-        })
+        }
 
         return nearest
     }
@@ -578,17 +472,17 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
 
     chooseEdge(position) {
-        if (!this.props.stories.active) {
+        if (this.props.stories.active === null) {
             return null
         }
 
+        const activeStory = AStorytelling.getActive(this.props.stories)
+
         position = CameraTransformations.screenToWorld(position, this.createTransform())
 
-        for (let i = 0; i < this.props.stories.active.edges.length; i++) {
-            const edge = this.props.stories.active.edges[i]
-
-            const a = edge.source.getCenterAsVector2()
-            const b = edge.destination.getCenterAsVector2()
+        for (const edge of Object.values(activeStory.edges.byId)) {
+            const a = ACluster.getCenterAsVector2(this.props.workspace, AStorytelling.retrieveCluster(this.props.stories, edge.source))
+            const b = ACluster.getCenterAsVector2(this.props.workspace, AStorytelling.retrieveCluster(this.props.stories, edge.destination))
             const dir = b.clone().sub(a.clone()).normalize()
             const l = new THREE.Vector2(-dir.y, dir.x).multiplyScalar(0.5)
             const r = new THREE.Vector2(dir.y, -dir.x).multiplyScalar(0.5)
@@ -616,10 +510,10 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         var res = -1
 
         for (var index = 0; index < this.props.dataset.vectors.length; index++) {
-            var value = this.props.dataset.vectors[index]
+            var value = this.props.workspace[index]
 
             // Skip points matching some criteria
-            if (!this.particles.isPointVisible(value)) {
+            if (!this.particles.isPointVisible(this.props.dataset.vectors[index])) {
                 continue
             }
 
@@ -649,6 +543,7 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
     }
 
     componentDidMount() {
+        this.initializeContainerEvents()
         this.setupRenderer()
         this.startRendering()
     }
@@ -686,22 +581,26 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
     onMouseLeave(event) {
         event.preventDefault()
-        this.mouseController.mouseLeave(event)
+        if (this.props.dataset)
+            this.mouseController.mouseLeave(event)
     }
 
     onMouseDown(event) {
         event.preventDefault()
-        this.mouseController.mouseDown(event)
+        if (this.props.dataset)
+            this.mouseController.mouseDown(event)
     }
 
     onMouseMove(event) {
         event.preventDefault();
-        this.mouseController.mouseMove(event)
+        if (this.props.dataset)
+            this.mouseController.mouseMove(event)
     }
 
     onMouseUp(event: MouseEvent) {
         event.preventDefault()
-        this.mouseController.mouseUp(event)
+        if (this.props.dataset)
+            this.mouseController.mouseUp(event)
     }
 
 
@@ -717,10 +616,10 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         }
 
         this.props.dataset.vectors.forEach((vector, index) => {
-            vector.view.selected = false
+            vector.__meta__.selected = false
         })
 
-        this.props.setCurrentAggregation([])
+        this.props.selectVectors([], false)
 
         this.particles?.updateColor()
     }
@@ -728,6 +627,10 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
     onWheel(event) {
         event.preventDefault()
+
+        if (!this.props.dataset) {
+            return;
+        }
 
         var normalized = normalizeWheel(event)
 
@@ -780,14 +683,15 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
     setupRenderer() {
         this.renderer = new THREE.WebGLRenderer({
-            antialias: true
+            antialias: true,
+            alpha: true
         });
         this.renderer.autoClear = true
         this.renderer.autoClearColor = false
 
         this.renderer.setPixelRatio(window.devicePixelRatio);
         this.renderer.setSize(this.getWidth(), this.getHeight());
-        this.renderer.setClearColor(0xf9f9f9, 1);
+        this.renderer.setClearColor(0x000000, 0);
         this.renderer.sortObjects = false;
 
         this.camera = new THREE.OrthographicCamera(this.getWidth() / - 2, this.getWidth() / 2, this.getHeight() / 2, this.getHeight() / - 2, 1, 1000);
@@ -798,9 +702,6 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
         this.containerRef.current.appendChild(this.renderer.domElement);
 
-
-        this.scene = new THREE.Scene()
-
         this.props.setViewTransform(this.camera, this.getWidth(), this.getHeight())
 
         this.prevTime = performance.now()
@@ -808,6 +709,8 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
     }
 
     createVisualization(dataset, lineColorScheme, vectorColorScheme) {
+        this.disposeScene()
+
         this.scene = new THREE.Scene()
         this.pointScene = new THREE.Scene()
         this.segments = dataset.segments
@@ -835,16 +738,21 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
             this.lines.meshes.forEach(line => this.scene.add(line.line))
         }
 
-        this.particles = new PointVisualization(this.vectorColorScheme, this.props.dataset, window.devicePixelRatio * 8, this.lines?.grayedLayerSystem)
-        this.particles.createMesh(this.props.dataset.vectors, this.segments)
+        this.particles = new PointVisualization(this.vectorColorScheme, this.props.dataset, window.devicePixelRatio * 8, this.lines?.grayedLayerSystem, this.segments)
+        this.particles.createMesh(this.props.dataset.vectors, this.segments, () => { this.requestRender() })
         this.particles.zoom(this.camera.zoom)
         this.particles.update()
 
 
         //this.scene.add(this.particles.mesh);
         this.pointScene.add(this.particles.mesh)
+    }
 
 
+
+
+
+    initializeContainerEvents() {
         var container = this.containerRef.current
         // Remove old listeners
         container.removeEventListener('mousemove', this.mouseMoveListener)
@@ -874,15 +782,19 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
     }
 
 
+
+
+
+
     filterLines(algo, show) {
 
         this.segments.forEach((segment) => {
             if (segment.vectors[0].algo == algo) {
-                segment.view.globalVisible = show
+                segment.__meta__.globalVisible = show
 
 
                 segment.vectors.forEach((vector) => {
-                    vector.view.visible = show
+                    vector.__meta__.visible = show
                 })
             }
         })
@@ -900,7 +812,7 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
     setLineFilter(checked) {
         this.segments.forEach((segment) => {
             var show = checked[segment.vectors[0].line]
-            segment.view.detailVisible = show
+            segment.__meta__.detailVisible = show
         })
         this.lines.update()
         this.particles.update()
@@ -912,19 +824,25 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
      * recalculate the optimal camera zoom level.
      */
     updateXY() {
-        this.particles.updatePosition()
-        if (this.props.dataset.isSequential) {
-            this.lines.updatePosition()
+        if (this.props.workspace) {
+            this.particles.updatePosition(this.props.workspace)
+
+            if (this.props.dataset.isSequential) {
+                this.lines.updatePosition(this.props.workspace)
+            }
+
+            ADataset.calculateBounds(this.props.dataset)
+
+
+            this.camera.zoom = getDefaultZoom(this.props.dataset.vectors, this.getWidth(), this.getHeight())
+            this.camera.position.x = 0.0
+            this.camera.position.y = 0.0
+            this.camera.updateProjectionMatrix();
+
+            this.repositionClusters()
+
+            this.requestRender()
         }
-        this.props.dataset.calculateBounds()
-        this.camera.zoom = getDefaultZoom(this.props.dataset.vectors, this.getWidth(), this.getHeight())
-        this.camera.position.x = 0.0
-        this.camera.position.y = 0.0
-        this.camera.updateProjectionMatrix();
-
-
-
-        this.requestRender()
     }
 
 
@@ -958,16 +876,12 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
     disposeScene() {
         this.currentHover = null
 
-        this.setState({
-            displayClusters: []
-        })
-
-        if (this.lines != null) {
+        if (this.lines != null && this.scene != null) {
             this.lines.dispose(this.scene)
             this.lines = null
         }
 
-        if (this.particles != null) {
+        if (this.particles != null && this.scene != null) {
             this.scene.remove(this.particles.mesh)
             this.particles.dispose()
             this.particles = null
@@ -1035,7 +949,7 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         this.renderLasso(ctx)
 
         if (this.clusterDrag) {
-            this.clusterDrag.renderToContext(extended, CameraTransformations.worldToScreen(this.clusterDrag.cluster.getCenter(), this.createTransform()), this.mouseController.currentMousePosition)
+            this.clusterDrag.renderToContext(extended, CameraTransformations.worldToScreen(ACluster.getCenterFromWorkspace(this.props.workspace, this.clusterDrag.cluster), this.createTransform()), this.mouseController.currentMousePosition)
         }
 
         if (this.traceSelect) {
@@ -1049,8 +963,9 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         }
     }
 
-    onClusterClicked(cluster: Cluster, shiftKey: boolean = false) {
-        this.props.setSelectedCluster(cluster, shiftKey)
+    onClusterClicked(cluster: ICluster, shiftKey: boolean = false) {
+        const activeStory = AStorytelling.getActive(this.props.stories)
+        this.props.setSelectedCluster([Object.keys(activeStory.clusters.byId).find(key => activeStory.clusters.byId[key] === cluster)], shiftKey)
     }
 
 
@@ -1102,7 +1017,7 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         }
     }
 
-    private updateItemClusterDisplay() {
+    updateItemClusterDisplay() {
         switch (this.props.displayMode) {
             case DisplayMode.StatesAndClusters:
             case DisplayMode.OnlyStates:
@@ -1129,10 +1044,16 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         }
     }
 
-    componentDidUpdate(prevProps, prevState) {
-        if (prevProps.pointColorScale != this.props.pointColorScale || prevProps.stories != this.props.stories) {
-            if (this.props.channelColor && this.props.pointColorScale) {
-                let mapping = mappingFromScale(this.props.pointColorScale, this.props.channelColor, this.props.dataset)
+    componentDidUpdate(prevProps: Props, prevState) {
+        if (prevProps.dataset !== this.props.dataset) {
+            this.createVisualization(this.props.dataset, mappingFromScale({ type: 'categorical', palette: 'dark2' }, { key: 'algo' }, this.props.dataset), null)
+        }
+
+        if (prevProps.colorScales.active != this.props.colorScales.active || prevProps.stories != this.props.stories) {
+
+            if (this.props.channelColor && this.props.colorScales.active) {
+
+                let mapping = mappingFromScale(ANormalized.get(this.props.colorScales.scales, this.props.colorScales.active), this.props.channelColor, this.props.dataset)
                 this.props.setPointColorMapping(mapping)
                 this.particles.colorCat(this.props.channelColor, mapping)
             } else {
@@ -1142,10 +1063,10 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         }
 
         if (prevProps.stories != this.props.stories) {
-            if (this.props.stories.active) {
+            if (this.props.stories.active !== null) {
                 this.particles?.storyTelling(this.props.stories)
 
-                this.lines?.storyTelling(this.props.stories)
+                this.lines?.storyTelling(this.props.stories, this.props.dataset.vectors)
                 this.lines?.update()
 
                 this.particles?.update()
@@ -1153,7 +1074,7 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
             } else {
                 this.particles?.storyTelling(null)
 
-                this.lines?.storyTelling(this.props.stories)
+                this.lines?.storyTelling(this.props.stories, this.props.dataset.vectors)
                 this.lines?.update()
 
                 this.particles?.update()
@@ -1183,17 +1104,19 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
 
         if (!arraysEqual(prevProps.currentAggregation.aggregation, this.props.currentAggregation.aggregation)) {
+            const vectorSelection = this.props.currentAggregation.aggregation.map(i => this.props.dataset.vectors[i])
+
             if (this.props.dataset.isSequential) {
-                var uniqueIndices = [...new Set(this.props.currentAggregation.aggregation.map(vector => vector.view.segment.lineKey))]
+                var uniqueIndices = [...new Set(vectorSelection.map(vector => vector.line))]
 
                 this.lines.groupHighlight(uniqueIndices)
             } else {
                 this.particles.groupHighlight(this.props.currentAggregation.aggregation)
             }
 
-            this.props.dataset.vectors.forEach(sample => sample.view.selected = false)
-            this.props.currentAggregation.aggregation.forEach(sample => {
-                sample.view.selected = true
+            this.props.dataset.vectors.forEach(sample => sample.__meta__.selected = false)
+            vectorSelection.forEach(sample => {
+                sample.__meta__.selected = true
             })
 
             this.particles.update()
@@ -1201,15 +1124,41 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         }
 
 
+        if (prevProps.hoverState !== this.props.hoverState) {
+            const hover_item = this.props.hoverState.data
+
+            if (isVector(hover_item)) {
+                let idx = hover_item.__meta__.meshIndex;
+
+                this.particles.highlight(idx);
+                if (this.props.dataset.isSequential) {
+                    if (idx >= 0) {
+                        this.lines.highlight([this.props.dataset.vectors[idx].line], this.getWidth(), this.getHeight(), this.scene)
+                    } else {
+                        this.lines.highlight([], this.getWidth(), this.getHeight(), this.scene)
+                    }
+                }
+                if (idx >= 0) {
+                    if (this.currentHover != this.props.dataset.vectors[idx]) {
+                        this.currentHover = this.props.dataset.vectors[idx]
+                        this.requestRender()
+                    }
+                } else {
+                    if (this.currentHover) {
+                        this.currentHover = null
+                        this.requestRender()
+                    }
+                }
+            }
 
 
+        }
 
         // Path length range has changed, update view accordingly
         if (this.props.dataset && this.props.dataset.isSequential && prevProps.pathLengthRange !== this.props.pathLengthRange) {
             // Set path length range on all segment views, and update them
-            this.segments.forEach(segment => {
-                segment.view.pathLengthRange = this.props.pathLengthRange.range
-            })
+            this.lines.pathLengthRange = this.props.pathLengthRange.range
+            this.particles.pathLengthRange = this.props.pathLengthRange.range
 
             this.lines.update()
             this.particles.update()
@@ -1221,13 +1170,13 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
             this.particles.shapeCat(this.props.vectorByShape)
         }
 
-        if (prevProps.checkedShapes != this.props.checkedShapes && this.particles) {
-            this.filterPoints(this.props.checkedShapes)
+        if (prevProps.pointDisplay.checkedShapes != this.props.pointDisplay.checkedShapes && this.particles) {
+            this.filterPoints(this.props.pointDisplay.checkedShapes)
         }
 
         if (prevProps.activeLine != this.props.activeLine) {
             if (this.props.activeLine != null) {
-                this.lines.highlight([this.props.activeLine.lineKey], this.getWidth(), this.getHeight(), this.scene, true)
+                this.lines.highlight([this.props.activeLine], this.getWidth(), this.getHeight(), this.scene, true)
             } else {
                 this.lines.highlight([], this.getWidth(), this.getHeight(), this.scene, true)
                 this.particles.update()
@@ -1237,6 +1186,10 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
         if (prevProps.advancedColoringSelection != this.props.advancedColoringSelection && this.particles) {
             this.particles.colorFilter(this.props.advancedColoringSelection)
+        }
+
+        if (prevProps.workspace !== this.props.workspace) {
+            this.updateXY()
         }
 
         this.requestRender()
@@ -1250,11 +1203,13 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         requestAnimationFrame(() => this.renderFrame())
     }
 
-    createTransform() {
+    createTransform(): ViewTransformType {
         return {
-            camera: this.camera,
+            centerX: this.camera.position.x,
+            centerY: this.camera.position.y,
             width: this.getWidth(),
-            height: this.getHeight()
+            height: this.getHeight(),
+            zoom: this.camera.zoom
         }
     }
 
@@ -1300,22 +1255,9 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
         this.multivariateClusterView?.current.createTriangulatedMesh()
     }
 
-    loadProjection(projection: Embedding) {
-        this.props.dataset.vectors.forEach(vector => {
-            let position = projection.positions[vector.view.meshIndex]
-            vector.x = position.x
-            vector.y = position.y
-        })
-
-        this.updateXY()
-    }
-
     onClusterZoom(cluster) {
         this.props.setSelectedCluster(cluster, false)
-        //this.props.setCurrentAggregation(cluster.vectors)
-        //this.props.setSelectedClusters([cluster])
     }
-
 
     render() {
         const handleClose = () => {
@@ -1336,9 +1278,17 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
                 width: '100%',
                 height: '100%',
                 overflow: "hidden",
-                position: "relative",
-                cursor: getToolCursor(this.props.currentTool)
+                position: "relative"
             }}>
+
+
+            {
+                // Layers which are behind the webgl view
+                this.props.overrideComponents?.layers?.filter(layer => layer.order < 0).map(layer => {
+                    return React.isValidElement(layer.component) ? layer.component : React.createElement(layer.component as () => JSX.Element)
+                })
+            }
+
 
             <div id="container" style={{
                 position: "absolute",
@@ -1346,17 +1296,22 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
                 left: "0px",
                 width: "100%",
                 height: "100%"
-            }} ref={this.containerRef} tabIndex={0}>
+            }} ref={this.containerRef} tabIndex={0}></div>
 
-
-
-
-            </div>
 
             <LassoLayer ref={this.selectionRef}></LassoLayer>
 
-            <MultivariateClustering ref={this.multivariateClusterView}></MultivariateClustering>
+            <MultivariateClustering
+                onInvalidate={() => { this.requestRender() }}
+                ref={this.multivariateClusterView}></MultivariateClustering>
 
+
+            {
+                // Layers which are in front of the webgl view
+                this.props.overrideComponents?.layers?.filter(layer => layer.order > 0).map(layer => {
+                    return React.isValidElement(layer.component) ? layer.component : React.createElement(layer.component as () => JSX.Element)
+                })
+            }
 
 
 
@@ -1373,12 +1328,24 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
             >
                 <MenuItem onClick={() => {
                     if (this.props.currentAggregation.aggregation.length > 0) {
-                        let cluster = Cluster.fromSamples(this.props.currentAggregation.aggregation)
+                        let cluster = ACluster.fromSamples(this.props.dataset, this.props.currentAggregation.aggregation)
 
-                        if (!this.props.stories.active) {
-                            let story = new Storybook([cluster], [])
-                            this.props.addStory(story)
-                            this.props.setActiveStory(story)
+                        if (this.props.stories.active === null) {
+                            const handle = uuidv4()
+                            const story: IBook = {
+                                clusters: {
+                                    byId: {
+                                        [handle]: cluster
+                                    },
+                                    allIds: [handle]
+                                },
+                                edges: {
+                                    byId: {},
+                                    allIds: []
+                                }
+                            }
+
+                            this.props.addStory(story, true)
                         } else {
                             this.props.addClusterToStory(cluster)
                         }
@@ -1397,7 +1364,7 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
                     var idx = this.choose(coords)
                     if (idx >= 0) {
                         var vector = this.props.dataset.vectors[idx]
-                        this.props.setActiveLine(vector.view.segment)
+                        this.props.setActiveLine(vector.line)
                     }
                 }}>Investigate Line</MenuItem>
             </Menu>
@@ -1406,7 +1373,7 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
             <Menu
                 keepMounted
-                open={this.state.menuY !== null && this.state.menuTarget instanceof Cluster}
+                open={this.state.menuY !== null && isCluster(this.state.menuTarget)}
                 onClose={handleClose}
                 anchorReference="anchorPosition"
                 anchorPosition={
@@ -1416,29 +1383,37 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
                 }
             >
                 <MenuItem onClick={() => {
-                    this.props.removeClusterFromStories(this.state.menuTarget)
-
+                    if (isCluster(this.state.menuTarget)) {
+                        this.props.removeClusterFromStories(this.state.menuTarget)
+                    }
 
                     handleClose()
                 }}>{'Delete Group'}</MenuItem>
 
                 <MenuItem onClick={() => {
-                    let paths = this.props.stories.active.getAllStoriesFromSource(this.state.menuTarget.label)
+                    if (!isCluster(this.state.menuTarget)) {
+                        handleClose()
+                        return;
+                    }
 
+                    const activeStory = AStorytelling.getActive(this.props.stories)
 
+                    let paths = ABook.getAllStoriesFromSource(activeStory, Object.entries(activeStory.clusters.byId).find(([key, val]) => val === this.state.menuTarget)[0])
 
                     if (paths.length > 0) {
-                        let mainPath = paths[0].map(id => this.props.stories.active.clusters.find(e => e.label == id))
+                        let mainPath = paths[0]
                         let mainEdges = mainPath.slice(1).map((item, index) => {
-                            return this.props.stories.active.edges.find(edge => edge.source == mainPath[index] && edge.destination == item)
+                            const [resultEdgeKey, _] = Object.entries(activeStory.edges.byId).find(([key, edge]) => edge.source === mainPath[index] && edge.destination === item)
+                            return resultEdgeKey
                         })
                         this.props.setActiveTrace({
                             mainPath: mainPath,
                             mainEdges: mainEdges,
                             sidePaths: paths.slice(1).map(ids => {
-                                let path = ids.map(id => this.props.stories.active.clusters.find(e => e.label == id))
+                                let path = ids
                                 let edges = path.slice(1).map((item, index) => {
-                                    return this.props.stories.active.edges.find(edge => edge.source == path[index] && edge.destination == item)
+                                    const [resultEdgeKey, _] = Object.entries(activeStory.edges.byId).find(([key, edge]) => edge.source === path[index] && edge.destination === item)
+                                    return resultEdgeKey
                                 })
                                 return {
                                     nodes: path,
@@ -1453,8 +1428,12 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
                 }}>{"Stories ... Starting from this Group"}</MenuItem>
 
                 <MenuItem onClick={() => {
-                    this.traceSelect = new TraceSelectTool(this.state.menuTarget)
+                    if (!isCluster(this.state.menuTarget)) {
+                        handleClose()
+                        return;
+                    }
 
+                    this.traceSelect = new TraceSelectTool(this.props.workspace, this.state.menuTarget)
                     handleClose()
                 }}>{"Stories ... Between 2 Groups"}</MenuItem>
 
@@ -1466,7 +1445,7 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
 
             <Menu
                 keepMounted
-                open={this.state.menuY !== null && this.state.menuTarget instanceof Edge}
+                open={this.state.menuY !== null && isEdge(this.state.menuTarget)}
                 onClose={handleClose}
                 anchorReference="anchorPosition"
                 anchorPosition={
@@ -1481,7 +1460,6 @@ export const WebGLView = connector(class extends React.Component<Props, ViewStat
                     handleClose()
                 }}>{'Delete Edge'}</MenuItem>
             </Menu>
-
         </div>
     }
 })
