@@ -1,3 +1,5 @@
+import { DiscreteMapping, ContinuousMapping } from "..";
+
 const SET = "ducks/pointColorMapping/SET"
 
 export const setPointColorMapping = pointColorMapping => ({
@@ -6,7 +8,7 @@ export const setPointColorMapping = pointColorMapping => ({
 });
 
 
-const pointColorMapping = (state = null, action) => {
+const pointColorMapping = (state = null, action): DiscreteMapping | ContinuousMapping => {
     switch (action.type) {
         case SET:
             return action.pointColorMapping
