@@ -1,137 +1,239 @@
 import { getSyncNodesAlt } from "../NumTs/NumTs";
-import Cluster from "../Utility/Data/Cluster";
-import { Edge } from "../Utility/graphs";
-import { Storybook } from "../Utility/Data/Storybook";
-import { Vect } from "../Utility/Data/Vect";
+import { ACluster, ICluster } from "../../model/Cluster";
+import { IEdge } from "../../model/Edge";
+import { IBook, ABook } from "../../model/Book";
+import { ObjectTypes } from "../../model/ObjectType";
+import { ANormalized } from "../Utility/NormalizedState";
+import { RootState } from "../Store";
 
-const ADD_STORY_BOOK = "ducks/stories/ADD"
-const DELETE = "ducks/stories/DELETE"
-const SET = "ducks/stories/SET"
-const ADD_CLUSTER_TO_ACTIVE = "ducks/stories/ADD_CLUSTER"
-const SET_ACTIVE_STORY_BOOK = "ducks/stories/SET_ACTIVE"
-const REMOVE_CLUSTER_FROM_STORIES = "ducks/stories/REMOVE_CLUSTER_FROM_STORIES"
-const ADD_EDGE_TO_ACTIVE = "ducks/stories/ADD_EDGE_TO_ACTIVE"
-const SET_ACTIVE_TRACE = "ducks/stories/SET_ACTIVE_TRACE"
-const ADD_CLUSTER_TO_TRACE = "ducks/stories/ADD_CLUSTER_TO_TRACE"
-const SET_ACTIVE_TRACE_STATE = "ducks/stories/SET_ACTIVE_TRACE_STATE"
-const SELECT_SIDE_BRANCH = "ducks/stories/SELECT_SIDE_BRANCH"
-const SET_VECTORS = "ducks/stories/SET_VECTORS"
-const REMOVE_EDGE_FROM_ACTIVE = "ducks/stories/REMOVE_EDGE_FROM_ACTIVE"
 
-export const addStory = story => ({
-    type: ADD_STORY_BOOK,
-    story: story
-});
+const enum ActionTypes {
+    ADD_BOOK = "ducks/stories/ADD",
+    DELETE_BOOK = "ducks/stories/DELETE",
 
-export const deleteStory = story => ({
-    type: DELETE,
-    story: story
-});
+    ADD_CLUSTER = "ducks/stories/ADD_CLUSTER",
+    DELETE_CLUSTER = "ducks/stories/REMOVE_CLUSTER_FROM_STORIES",
 
-export function setStories(stories: Storybook[]) {
-    return {
-        type: SET,
-        stories: stories
+    SET = "ducks/stories/SET",
+
+    SET_ACTIVE_STORY_BOOK = "ducks/stories/SET_ACTIVE",
+
+    ADD_EDGE_TO_ACTIVE = "ducks/stories/ADD_EDGE_TO_ACTIVE",
+    SET_ACTIVE_TRACE = "ducks/stories/SET_ACTIVE_TRACE",
+    ADD_CLUSTER_TO_TRACE = "ducks/stories/ADD_CLUSTER_TO_TRACE",
+    SET_ACTIVE_TRACE_STATE = "ducks/stories/SET_ACTIVE_TRACE_STATE",
+    SELECT_SIDE_BRANCH = "ducks/stories/SELECT_SIDE_BRANCH",
+    REMOVE_EDGE_FROM_ACTIVE = "ducks/stories/REMOVE_EDGE_FROM_ACTIVE"
+}
+
+
+/**type AddStoryAction = {
+    type: ActionTypes.ADD_STORY_BOOK
+    story: IBook
+    activate: boolean
+}**/
+
+//type Action = AddStoryAction
+
+export function addBook(story: IBook, activate: boolean = false) {
+    return (dispatch, getState) => {
+        const { dataset } = getState() as RootState
+
+        return dispatch({
+            type: ActionTypes.ADD_BOOK,
+            story: story,
+            activate: activate,
+            vectors: dataset.vectors
+        })
     }
 }
 
 
-export const addClusterToStory = cluster => ({
-    type: ADD_CLUSTER_TO_ACTIVE,
-    cluster: cluster
-})
+export function deleteBook(story: IBook) {
+    return (dispatch, getState) => {
+        const { dataset } = getState() as RootState
 
-export function setActiveStory(activeStory: Storybook) {
-    return {
-        type: SET_ACTIVE_STORY_BOOK,
-        activeStory: activeStory
+        return dispatch({
+            type: ActionTypes.DELETE_BOOK,
+            story: story,
+            vectors: dataset.vectors
+        })
     }
 }
 
-export function removeClusterFromStories(cluster: Cluster) {
-    return {
-        type: REMOVE_CLUSTER_FROM_STORIES,
-        cluster: cluster
+
+
+
+
+export function addCluster(cluster: ICluster) {
+    return (dispatch, getState) => {
+        const { dataset } = getState() as RootState
+
+        return dispatch({
+            type: ActionTypes.ADD_CLUSTER,
+            cluster: cluster,
+            vectors: dataset.vectors
+        })
     }
 }
+
+export function deleteCluster(cluster: ICluster) {
+    return (dispatch, getState) => {
+        const { dataset } = getState() as RootState
+
+        return dispatch({
+            type: ActionTypes.DELETE_CLUSTER,
+            cluster: cluster,
+            vectors: dataset.vectors
+        })
+    }
+}
+
+
+
+
+
+
+
+export function setStories(stories: IBook[]) {
+    return (dispatch, getState) => {
+        const { dataset } = getState() as RootState
+
+        return dispatch({
+            type: ActionTypes.SET,
+            stories: stories,
+            vectors: dataset.vectors
+        })
+    }
+}
+
+
+
+
+export function setActiveStory(activeStory: IBook) {
+    return (dispatch, getState) => {
+        const { dataset } = getState() as RootState
+
+        return dispatch({
+            type: ActionTypes.SET_ACTIVE_STORY_BOOK,
+            activeStory: activeStory,
+            vectors: dataset.vectors
+        })
+    }
+}
+
+
 
 export function addEdgeToActive(edge) {
     return {
-        type: ADD_EDGE_TO_ACTIVE,
+        type: ActionTypes.ADD_EDGE_TO_ACTIVE,
         edge: edge
     }
 }
 
 export function removeEdgeFromActive(edge) {
     return {
-        type: REMOVE_EDGE_FROM_ACTIVE,
+        type: ActionTypes.REMOVE_EDGE_FROM_ACTIVE,
         edge: edge
     }
 }
 
-export const setActiveTrace = activeTrace => ({
-    type: SET_ACTIVE_TRACE,
+export const setActiveTrace = (activeTrace: number) => ({
+    type: ActionTypes.SET_ACTIVE_TRACE,
     activeTrace: activeTrace
 })
 
-export const addClusterToTrace = cluster => ({
-    type: ADD_CLUSTER_TO_TRACE,
-    cluster: cluster
-})
+
+export const addClusterToTrace = (cluster) => {
+    return (dispatch, getState) => {
+        const { dataset } = getState() as RootState
+
+        return dispatch({
+            type: ActionTypes.ADD_CLUSTER_TO_TRACE,
+            cluster: cluster,
+            vectors: dataset.vectors
+        })
+    }
+}
 
 
-export function setActiveTraceState(cluster: Cluster) {
+export function setActiveTraceState(cluster: string) {
     return {
-        type: SET_ACTIVE_TRACE_STATE,
+        type: ActionTypes.SET_ACTIVE_TRACE_STATE,
         cluster: cluster
     }
 }
 
 export function selectSideBranch(i: number) {
     return {
-        type: SELECT_SIDE_BRANCH,
+        type: ActionTypes.SELECT_SIDE_BRANCH,
         index: i
     }
 }
 
-export function setVectors(vectors: Vect[]) {
-    return {
-        type: SET_VECTORS,
-        vectors: vectors
+
+export class AStorytelling {
+    static createEmpty(): IStorytelling {
+        return {
+            stories: [],
+            active: null,
+            trace: null,
+            activeTraceState: null
+        }
+    }
+
+    static emptyStory(): IBook {
+        const story: IBook = {
+            clusters: ANormalized.create<ICluster>(),
+            edges: ANormalized.create<IEdge>()
+        }
+
+        return story
+    }
+
+    static getActive(stories: IStorytelling): IBook {
+        if (stories && stories.stories) {
+            return stories.stories[stories.active]
+        } else {
+            return null
+        }
+    }
+
+    static retrieveCluster(stories: IStorytelling, clusterIndex: string): ICluster {
+        return stories.stories[stories.active].clusters.byId[clusterIndex]
+    }
+
+    static retreiveEdge(stories: IStorytelling, edgeIndex: string): IEdge {
+        return stories.stories[stories.active].edges.byId[edgeIndex]
     }
 }
 
-const initialState = {
-    vectors: [],
-    stories: [],
-    active: null,
-    trace: null,
-    activeTraceState: null
+const initialState = AStorytelling.createEmpty()
+
+/**
+ * Type interface for stories slace of the redux store.
+ */
+export type IStorytelling = {
+    stories: IBook[]
+
+    active: number
+
+    trace: { mainPath: string[], mainEdges: string[], sidePaths: { nodes: string[], edges: string[], syncNodes: number[] }[] }
+
+    activeTraceState: string
 }
 
-export type StoriesType = {
-    vectors: Vect[]
-
-    stories: Storybook[]
-
-    active: Storybook
-
-    trace: { mainPath: Cluster[], mainEdges: any[], sidePaths: { nodes: Cluster[], edges: Edge[], syncNodes: number[] }[] }
-
-    activeTraceState: Cluster
-}
-
-export default function stories(state: StoriesType = initialState, action): StoriesType {
+export default function stories(state: IStorytelling = initialState, action): IStorytelling {
     switch (action.type) {
-        case SET_VECTORS:
-            let clone = Object.assign({}, state)
-            clone.vectors = action.vectors
-            return clone
-        case SELECT_SIDE_BRANCH: {
+        case ActionTypes.SELECT_SIDE_BRANCH: {
             let sidePaths = state.trace.sidePaths.slice(0)
 
             sidePaths.splice(action.index, 1)
-            sidePaths.push({ nodes: state.trace.mainPath, edges: state.trace.mainEdges, syncNodes: [] })
+            sidePaths.push({
+                nodes: state.trace.mainPath,
+                edges: state.trace.mainEdges,
+                syncNodes: []
+            })
 
             let trace = {
                 mainPath: state.trace.sidePaths[action.index].nodes,
@@ -144,78 +246,81 @@ export default function stories(state: StoriesType = initialState, action): Stor
             })
 
             return {
-                vectors: state.vectors,
                 stories: state.stories,
                 active: state.active,
                 trace: trace,
                 activeTraceState: state.activeTraceState
             }
         }
-        case SET_ACTIVE_TRACE_STATE: {
+        case ActionTypes.SET_ACTIVE_TRACE_STATE: {
+
             return {
-                vectors: state.vectors,
                 stories: state.stories,
                 active: state.active,
                 trace: state.trace,
                 activeTraceState: action.cluster
             }
         }
-        case ADD_CLUSTER_TO_TRACE: {
+        case ActionTypes.ADD_CLUSTER_TO_TRACE: {
             let cluster = action.cluster
 
             // Add cluster to active story book
-            state.active.clusters.push(cluster)
+            const activeStory = AStorytelling.getActive(state)
+
+            ABook.addCluster(activeStory, cluster)
 
             // Add edge that connects the active trace state with the current cluster
             if (state.trace.mainPath.length > 0) {
-                let edge = new Edge(state.trace.mainPath[state.trace.mainPath.length - 1], cluster, null)
-                state.active.edges.push(edge)
-                state.trace.mainEdges.push(edge)
+                let edge: IEdge = {
+                    source: state.trace.mainPath[state.trace.mainPath.length - 1],
+                    destination: cluster,
+                    objectType: ObjectTypes.Edge
+                }
+
+                const handle = ABook.addEdge(activeStory, edge)
+
+                state.trace.mainEdges.push(handle)
             }
 
             // Add cluster to current trace
             state.trace.mainPath.push(cluster)
 
-            Cluster.deriveVectorLabelsFromClusters(state.vectors, state.active.clusters)
+            ACluster.deriveVectorLabelsFromClusters(action.vectors, Object.values(activeStory.clusters.byId))
 
             return {
-                vectors: state.vectors,
                 stories: state.stories,
                 active: state.active,
                 trace: state.trace,
                 activeTraceState: state.activeTraceState
             }
         }
-        case SET_ACTIVE_TRACE: {
+        case ActionTypes.SET_ACTIVE_TRACE: {
             return {
-                vectors: state.vectors,
                 stories: state.stories,
                 active: state.active,
                 trace: action.activeTrace,
                 activeTraceState: state.activeTraceState
             }
         }
-        case SET:
-            if (state.active) {
-                Cluster.deriveVectorLabelsFromClusters(state.vectors, state.active.clusters)
+        case ActionTypes.SET:
+            if (state.active !== null) {
+                ACluster.deriveVectorLabelsFromClusters(action.vectors, Object.values(AStorytelling.getActive(state).clusters.byId))
             }
 
             return {
-                vectors: state.vectors,
                 stories: action.stories,
                 active: state.active,
                 trace: state.trace,
                 activeTraceState: state.activeTraceState
             }
-        case DELETE:
+        case ActionTypes.DELETE_BOOK:
             const newState = state.stories.slice(0)
             newState.splice(newState.indexOf(action.story), 1)
 
             if (state.active == action.story) {
-                Cluster.deriveVectorLabelsFromClusters(state.vectors, [])
+                ACluster.deriveVectorLabelsFromClusters(action.vectors, [])
 
                 return {
-                    vectors: state.vectors,
                     stories: newState,
                     active: null,
                     trace: null,
@@ -224,18 +329,18 @@ export default function stories(state: StoriesType = initialState, action): Stor
             }
 
             return {
-                vectors: state.vectors,
                 stories: newState,
                 active: state.active,
                 trace: state.trace,
                 activeTraceState: state.activeTraceState
             }
-        case ADD_STORY_BOOK: {
-            let storyBook = action.story as Storybook
+        case ActionTypes.ADD_BOOK: {
+            let storyBook = action.story as IBook
             let trace = state.trace
             let activeTraceState = state.activeTraceState
 
-            if (storyBook.clusters.length == 0) {
+
+            if (storyBook.clusters.allIds.length == 0) {
                 trace = {
                     mainEdges: [],
                     mainPath: [],
@@ -244,35 +349,34 @@ export default function stories(state: StoriesType = initialState, action): Stor
                 activeTraceState = null
             }
 
-            Cluster.deriveVectorLabelsFromClusters(state.vectors, storyBook.clusters)
+            ACluster.deriveVectorLabelsFromClusters(action.vectors, Object.values(storyBook.clusters.byId))
 
             if (state && state.stories.length > 0) {
                 const newState = state.stories.slice(0)
-                newState.push(storyBook)
+                const i = newState.push(storyBook) - 1
 
                 return {
-                    vectors: state.vectors,
                     stories: newState,
-                    active: storyBook,
+                    active: action.activate ? i : state.active,
                     trace: trace,
                     activeTraceState: activeTraceState
                 }
             } else {
                 return {
-                    vectors: state.vectors,
                     stories: [storyBook],
-                    active: storyBook,
+                    active: 0,
                     trace: trace,
                     activeTraceState: activeTraceState
                 }
             }
         }
-        case SET_ACTIVE_STORY_BOOK: {
-            let storyBook = action.activeStory as Storybook
+        case ActionTypes.SET_ACTIVE_STORY_BOOK: {
+            let storyBook = state.stories[action.activeStory]
 
             let trace = state.trace
             let activeTraceState = state.activeTraceState
-            if (storyBook && storyBook.clusters.length == 0) {
+
+            if (storyBook && storyBook.clusters.allIds.length == 0) {
                 trace = {
                     mainPath: [],
                     mainEdges: [],
@@ -284,38 +388,36 @@ export default function stories(state: StoriesType = initialState, action): Stor
             }
 
             if (storyBook && storyBook.clusters) {
-                Cluster.deriveVectorLabelsFromClusters(state.vectors, storyBook.clusters)
+                ACluster.deriveVectorLabelsFromClusters(action.vectors, Object.values(storyBook.clusters.byId))
             } else {
-                Cluster.deriveVectorLabelsFromClusters(state.vectors, [])
+                ACluster.deriveVectorLabelsFromClusters(action.vectors, [])
             }
-            
+
             return {
-                vectors: state.vectors,
                 stories: state.stories,
                 active: action.activeStory,
                 trace: trace,
                 activeTraceState: activeTraceState
             }
         }
-        case REMOVE_CLUSTER_FROM_STORIES: {
-            let cluster = action.cluster
+        case ActionTypes.DELETE_CLUSTER: {
+            const cluster = action.cluster as ICluster
 
-            // Find stories where the cluster is located
-            state.stories.forEach(story => {
-                if (story.clusters.includes(cluster)) {
-                    story.clusters.splice(story.clusters.indexOf(cluster), 1)
+            const activeStory = AStorytelling.getActive(state)
 
-                    story.edges.filter(edge => {
-                        return edge.source == cluster || edge.destination == cluster
-                    }).forEach(edge => {
-                        story.edges.splice(story.edges.indexOf(edge), 1)
-                    })
-                }
+            const handle = ABook.deleteCluster(activeStory, cluster)
+
+            const entries = Object.entries(activeStory.edges.byId).filter(([edgeHandle, edge]) => {
+                return edge.source === handle || edge.destination === handle
             })
+
+            for (const [handle, edge] of entries) {
+                ABook.deleteEdge(activeStory, edge)
+            }
 
             // Remove cluster labels from samples
             // TODO: check if this is ok in a reducer
-            cluster.vectors.forEach(sample => {
+            cluster.indices.map(i => action.vectors[i]).forEach(sample => {
                 if (Array.isArray(sample.groupLabel)) {
                     sample.groupLabel.splice(sample.groupLabel.indexOf(cluster.label), 1)
                 } else {
@@ -324,32 +426,35 @@ export default function stories(state: StoriesType = initialState, action): Stor
             })
 
             return {
-                vectors: state.vectors,
                 stories: state.stories,
                 active: state.active,
                 trace: state.trace,
                 activeTraceState: state.activeTraceState
             }
         }
-        case ADD_EDGE_TO_ACTIVE: {
-            state.active.edges.push(action.edge)
+        case ActionTypes.ADD_EDGE_TO_ACTIVE: {
+            const activeStory = AStorytelling.getActive(state)
+
+            ANormalized.add(activeStory.edges, action.edge)
 
             return {
-                vectors: state.vectors,
                 stories: state.stories,
                 active: state.active,
                 trace: state.trace,
                 activeTraceState: state.activeTraceState
             }
         }
-        case ADD_CLUSTER_TO_ACTIVE: {
-            let cluster = action.cluster
+        case ActionTypes.ADD_CLUSTER: {
+            let cluster = action.cluster as ICluster
 
-            state.active.clusters.push(cluster)
+            const story = AStorytelling.getActive(state)
+
+            ABook.addCluster(story, cluster)
 
             // Add cluster labels to samples
             // TODO: check if this is ok in a reducer
-            cluster.vectors.forEach(sample => {
+            cluster.indices.forEach(i => {
+                const sample = action.vectors[i]
                 if (Array.isArray(sample.groupLabel)) {
                     sample.groupLabel.push(cluster.label)
                 } else {
@@ -358,17 +463,17 @@ export default function stories(state: StoriesType = initialState, action): Stor
             })
 
             return {
-                vectors: state.vectors,
                 stories: state.stories,
                 active: state.active,
                 trace: state.trace,
                 activeTraceState: state.activeTraceState
             }
         }
-        case REMOVE_EDGE_FROM_ACTIVE: {
-            state.active.edges.splice(state.active.edges.indexOf(action.edge), 1)
+        case ActionTypes.REMOVE_EDGE_FROM_ACTIVE: {
+            const activeStory = AStorytelling.getActive(state)
+            ANormalized.deleteByRef(activeStory.edges, action.edge)
 
-            return { ... state }
+            return { ...state }
         }
         default:
             return state
