@@ -1,5 +1,5 @@
 import { isNumber } from "lodash";
-import { Vect } from "./Vect";
+import { IVector } from "../../../model/Vector";
 
 /**
  * Class that preprocesses the data set and checks for validity.
@@ -11,7 +11,7 @@ const DEFAULT_ALGO = "all"
 
 
 export class Preprocessor {
-    vectors: Vect[];
+    vectors: IVector[];
 
     constructor(vectors) {
         this.vectors = vectors;
@@ -159,7 +159,6 @@ export class Preprocessor {
         if (!header.includes('groupLabel')) {
             vectors.forEach(vector => {
                 vector.groupLabel = [];
-                vector.clusterProbability = 0.0;
             });
         } else {
             // Support multivariate points ... eg each groupLabel is actually an array
